+++ src/Request/MessageRequest/RequestTextMessage.php
@@ -11,7 +11,7 @@ final class RequestTextMessage extends MessageRequest
     */
     public function body(): array
     {
-        return [
+        $body = [
             'messaging_product' => $this->message->messagingProduct(),
             'recipient_type' => $this->message->recipientType(),
             'to' => $this->message->to(),
@@ -21,5 +21,11 @@ final class RequestTextMessage extends MessageRequest
                 'body' => $this->message->text(),
             ],
         ];
+
+        if ($this->message->replyTo()) {
+            $body['context']['message_id'] = $this->message->replyTo();
+        }
+
+        return $body;
     }
 }